+++ ts/packages/shell/src/renderer/src/chatView.ts
@@ -4,7 +4,6 @@
 import { IdGenerator, getClientAPI } from "./main";
 import { ChatInput, ExpandableTextarea, questionInput } from "./chatInput";
 import { SearchMenu } from "./search";
-import { AnsiUp } from "ansi_up";
 import { iconCheckMarkCircle, iconX, iconRoadrunner } from "./icon";
 import {
     ActionInfo,
@@ -16,17 +15,14 @@ import { ActionCascade } from "./ActionCascade";
 import {
     DisplayAppendMode,
     DisplayContent,
-    DisplayType,
-    DisplayMessageKind,
     DynamicDisplay,
 } from "@typeagent/agent-sdk";
-import { TTS, TTSMetrics } from "./tts/tts";
+import { TTS } from "./tts/tts";
 import { IAgentMessage } from "agent-dispatcher";
-import DOMPurify from "dompurify";
-import { PhaseTiming, RequestMetrics } from "agent-dispatcher";
 
 import { PartialCompletion } from "./partial";
-import { ChoicePanel, InputChoice } from "./choicePanel";
+import { InputChoice } from "./choicePanel";
+import { MessageGroup } from "./messageGroup";
 
 interface ISymbolNode {
     symbolName: string;
@@ -216,787 +212,6 @@ export class PlayerShimCursor {
     }
     symbolNames() {
         return this.cursor.children.map((child) => child.symbolName);
-    }
-}
-
-function metricsString(name: string, duration: number, count = 1) {
-    const avg = duration / count;
-    return `${name}: <b>${formatTimeReaderFriendly(avg)}${count !== 1 ? `(out of ${count})` : ""}</b>`;
-}
-
-class MessageContainer {
-    public readonly div: HTMLDivElement;
-    private readonly messageBodyDiv: HTMLDivElement;
-    private readonly messageDiv: HTMLDivElement;
-    private readonly timestampDiv: HTMLDivElement;
-    private readonly iconDiv: HTMLDivElement;
-
-    private metricsDiv?: {
-        mainMetricsDiv: HTMLDivElement;
-        markMetricsDiv: HTMLDivElement;
-        ttsMetricsDiv: HTMLDivElement;
-        firstResponseMetricsDiv: HTMLDivElement;
-    };
-
-    private messageStart?: number;
-    private audioStart?: number;
-    private ttsFirstChunkTotal: number = 0;
-    private ttsFirstChunkCount: number = 0;
-    private ttsSynthesisTotal: number = 0;
-
-    private lastAppendMode?: DisplayAppendMode;
-    private completed = false;
-    private pendingSpeakText: string = "";
-
-    public get source() {
-        return this._source;
-    }
-
-    private updateSource() {
-        const source = this._source;
-        const sourceIcon = this.agents.get(source);
-
-        // set source and source icon
-        (this.timestampDiv.firstChild as HTMLDivElement).innerText = source; // name
-        this.iconDiv.innerText = sourceIcon ?? "❔"; // icon
-    }
-
-    constructor(
-        private chatView: ChatView,
-        className: string,
-        private _source: string,
-        private readonly agents: Map<string, string>,
-        beforeElem: Element,
-        private hideMetrics: boolean,
-        private readonly requestStart: number,
-        private showFirstResponseMetrics = false,
-    ) {
-        const div = document.createElement("div");
-        div.className = className;
-
-        const timestampDiv = createTimestampDiv(
-            new Date(),
-            "chat-timestamp-left",
-        );
-        div.append(timestampDiv);
-        this.timestampDiv = timestampDiv;
-
-        const agentIconDiv = document.createElement("div");
-        agentIconDiv.className = "agent-icon";
-        div.append(agentIconDiv);
-        this.iconDiv = agentIconDiv;
-
-        const messageBodyDiv = document.createElement("div");
-        const bodyClass = this.hideMetrics
-            ? "chat-message-body-hide-metrics"
-            : "chat-message-body";
-        messageBodyDiv.className = `${bodyClass} chat-message-agent`;
-        div.append(messageBodyDiv);
-        this.messageBodyDiv = messageBodyDiv;
-
-        const message = document.createElement("div");
-        message.className = "chat-message-content";
-        messageBodyDiv.append(message);
-        this.messageDiv = message;
-
-        // The chat message list has the style flex-direction: column-reverse;
-        beforeElem.before(div);
-
-        this.div = div;
-        this.updateSource();
-    }
-
-    public getMessage() {
-        return this.messageDiv.innerText;
-    }
-
-    public setMessage(
-        content: DisplayContent,
-        source: string,
-        appendMode?: DisplayAppendMode, // default to not appending.
-    ) {
-        if (typeof content !== "string" && content.kind === "info") {
-            // Don't display info
-            return;
-        }
-
-        if (this.messageStart === undefined) {
-            // Don't count dispatcher status messages as first response.
-            if (source !== "dispatcher") {
-                this.messageStart = performance.now();
-                this.updateFirstResponseMetrics();
-            }
-        }
-
-        // Flush last temporary reset the lastAppendMode.
-        this.flushLastTemporary();
-
-        this._source = source;
-        this.updateSource();
-
-        const speakText = setContent(
-            this.messageDiv,
-            content,
-            appendMode === "inline" && this.lastAppendMode !== "inline"
-                ? "block"
-                : appendMode,
-        );
-
-        this.speak(speakText, appendMode);
-
-        this.lastAppendMode = appendMode;
-
-        this.updateDivState();
-    }
-
-    public addChoicePanel(
-        choices: InputChoice[],
-        onSelected: (choice: InputChoice) => void,
-    ) {
-        const choicePanel = new ChoicePanel(
-            this.messageDiv,
-            choices,
-            (choice: InputChoice) => {
-                choicePanel.remove();
-                onSelected(choice);
-            },
-        );
-    }
-
-    private speakText(tts: TTS, speakText: string) {
-        let cbAudioStart: (() => void) | undefined;
-        if (this.audioStart === undefined) {
-            this.audioStart = -1;
-            cbAudioStart = () => {
-                this.audioStart = performance.now();
-                this.updateFirstResponseMetrics();
-            };
-        }
-        const p = tts.speak(speakText, cbAudioStart);
-        p.then((timing) => {
-            if (timing) {
-                this.addTTSTiming(timing);
-            }
-        });
-    }
-
-    private speak(
-        speakText: string | undefined,
-        appendMode?: DisplayAppendMode,
-    ) {
-        const tts = this.chatView.tts;
-        if (tts === undefined) {
-            return;
-        }
-        if (speakText === undefined) {
-            // Flush the pending text.
-            this.flushPendingSpeak(tts);
-            return;
-        }
-
-        if (appendMode !== "inline") {
-            this.flushPendingSpeak(tts);
-            this.speakText(tts, speakText);
-            return;
-        }
-
-        this.pendingSpeakText += speakText;
-        const minSpeak = 10; // TODO: Adjust this value.
-        if (this.pendingSpeakText.length <= minSpeak) {
-            // Too short to speak.
-            return;
-        }
-        const segmenter = new Intl.Segmenter(navigator.language, {
-            granularity: "sentence",
-        });
-        const segments = Array.from(segmenter.segment(this.pendingSpeakText));
-
-        if (segments.length < 2) {
-            // No sentence boundary.
-            return;
-        }
-
-        // Try Keep the last segment and speak the rest.
-        const index = segments[segments.length - 1].index;
-        if (index <= minSpeak) {
-            // Too short to speak.
-            return;
-        }
-
-        const speakTextPartial = this.pendingSpeakText.slice(0, index);
-        this.pendingSpeakText = this.pendingSpeakText.slice(index);
-        this.speakText(tts, speakTextPartial);
-    }
-
-    private flushPendingSpeak(tts: TTS) {
-        // Flush the pending text.
-        if (this.pendingSpeakText) {
-            this.speakText(tts, this.pendingSpeakText);
-            this.pendingSpeakText = "";
-        }
-    }
-
-    public complete() {
-        this.completed = true;
-        if (this.chatView.tts) {
-            this.flushPendingSpeak(this.chatView.tts);
-        }
-        this.flushLastTemporary();
-        this.updateDivState();
-    }
-
-    private updateDivState() {
-        if (this.completed && !this.messageDiv.firstChild) {
-            this.hide();
-        } else {
-            this.show();
-        }
-    }
-
-    private flushLastTemporary() {
-        if (this.lastAppendMode === "temporary") {
-            this.messageDiv.lastChild?.remove();
-            this.lastAppendMode = undefined;
-        }
-    }
-
-    private ensureMetricsDiv() {
-        if (this.metricsDiv === undefined) {
-            const metricsContainer = document.createElement("div");
-            metricsContainer.className =
-                "chat-message-metrics chat-message-metrics-left";
-            this.messageBodyDiv.append(metricsContainer);
-
-            const metricsDetails = document.createElement("div");
-            metricsDetails.className = "metrics-details";
-            metricsContainer.append(metricsDetails);
-
-            const left = document.createElement("div");
-            metricsDetails.append(left);
-            const middle = document.createElement("div");
-            metricsDetails.append(middle);
-            const right = document.createElement("div");
-            metricsDetails.append(right);
-
-            const firstResponseMetricsDiv = document.createElement("div");
-            left.append(firstResponseMetricsDiv);
-
-            const markMetricsDiv = document.createElement("div");
-            middle.append(markMetricsDiv);
-
-            // Only show with dev UI.
-            const ttsMetricsDiv = document.createElement("div");
-            ttsMetricsDiv.className = "tts-metrics";
-            middle.append(ttsMetricsDiv);
-
-            const mainMetricsDiv = document.createElement("div");
-            right.append(mainMetricsDiv);
-
-            this.metricsDiv = {
-                mainMetricsDiv,
-                markMetricsDiv,
-                ttsMetricsDiv,
-                firstResponseMetricsDiv,
-            };
-        }
-        return this.metricsDiv;
-    }
-
-    public updateMainMetrics(metrics?: PhaseTiming, total?: number) {
-        if (metrics === undefined && total === undefined) {
-            return;
-        }
-        const metricsDiv = this.ensureMetricsDiv();
-        updateMetrics(
-            metricsDiv.mainMetricsDiv,
-            metricsDiv.markMetricsDiv,
-            "Action",
-            metrics,
-            total,
-        );
-    }
-
-    private updateFirstResponseMetrics() {
-        if (this.showFirstResponseMetrics) {
-            const messages: string[] = [];
-            if (this.messageStart !== undefined) {
-                messages.push(
-                    metricsString(
-                        "First Message",
-                        this.messageStart - this.requestStart,
-                    ),
-                );
-            }
-            if (this.audioStart !== undefined) {
-                messages.push(
-                    metricsString(
-                        "First Audio",
-                        this.audioStart - this.requestStart,
-                    ),
-                );
-            }
-            const div = this.ensureMetricsDiv().firstResponseMetricsDiv;
-            div.innerHTML = messages.join("<br>");
-        } else if (this.metricsDiv) {
-            this.metricsDiv.firstResponseMetricsDiv.innerHTML = "";
-        }
-    }
-
-    public setFirstResponseMetricsVisibility(visible: boolean) {
-        this.showFirstResponseMetrics = visible;
-        this.updateFirstResponseMetrics();
-    }
-
-    public addTTSTiming(timing: TTSMetrics) {
-        const messages: string[] = [];
-        if (timing.firstChunkTime) {
-            this.ttsFirstChunkTotal += timing.firstChunkTime;
-            this.ttsFirstChunkCount++;
-        }
-        this.ttsSynthesisTotal += timing.duration;
-
-        if (this.ttsFirstChunkCount !== 0) {
-            messages.push(
-                metricsString(
-                    "TTS First Chunk",
-                    this.ttsFirstChunkTotal,
-                    this.ttsFirstChunkCount,
-                ),
-            );
-        }
-        messages.push(metricsString("TTS Synthesis", this.ttsSynthesisTotal));
-        this.ensureMetricsDiv().ttsMetricsDiv.innerHTML = messages.join("<br>");
-    }
-
-    public show() {
-        this.div.classList.remove("chat-message-hidden");
-    }
-    public hide() {
-        this.div.classList.add("chat-message-hidden");
-    }
-    public scrollIntoView() {
-        this.div.scrollIntoView(false);
-    }
-
-    public setMetricsVisible(visible: boolean) {
-        this.hideMetrics = !visible;
-        updateMetricsVisibility(visible, this.messageBodyDiv);
-    }
-}
-
-function updateMetricsVisibility(visible: boolean, div: HTMLDivElement) {
-    const addClass = visible
-        ? "chat-message-body"
-        : "chat-message-body-hide-metrics";
-    const removeClass = visible
-        ? "chat-message-body-hide-metrics"
-        : "chat-message-body";
-    div.classList.remove(removeClass);
-    div.classList.add(addClass);
-}
-
-class MessageGroup {
-    public readonly userMessageContainer: HTMLDivElement;
-    public readonly userMessageBody: HTMLDivElement;
-    public readonly userMessage: HTMLDivElement;
-    public metricsDiv?: {
-        mainMetricsDiv: HTMLDivElement;
-        markMetricsDiv: HTMLDivElement;
-    };
-    private statusMessage: MessageContainer | undefined;
-    private readonly agentMessages: MessageContainer[] = [];
-    private readonly start: number = performance.now();
-    constructor(
-        private readonly chatView: ChatView,
-        request: DisplayContent,
-        container: HTMLDivElement,
-        requestPromise: Promise<RequestMetrics | undefined> | undefined,
-        timeStamp: Date,
-        public agents: Map<string, string>,
-        private hideMetrics: boolean,
-    ) {
-        const userMessageContainer = document.createElement("div");
-        userMessageContainer.className = "chat-message-right";
-
-        const timeStampDiv = createTimestampDiv(
-            timeStamp,
-            "chat-timestamp-right",
-        );
-        userMessageContainer.appendChild(timeStampDiv);
-
-        const userMessageBody = document.createElement("div");
-        const bodyClass = this.hideMetrics
-            ? "chat-message-body-hide-metrics"
-            : "chat-message-body";
-        userMessageBody.className = `${bodyClass} chat-message-user`;
-        userMessageContainer.appendChild(userMessageBody);
-
-        const userMessage = document.createElement("div");
-        userMessage.className = "chat-message-content";
-        userMessageBody.appendChild(userMessage);
-
-        setContent(userMessage, request);
-
-        if (container.firstChild) {
-            container.firstChild.before(userMessageContainer);
-
-            userMessageContainer.scrollIntoView(false);
-        } else {
-            container.append(userMessageContainer);
-        }
-
-        this.userMessageContainer = userMessageContainer;
-        this.userMessageBody = userMessageBody;
-        this.userMessage = userMessage;
-
-        this.chatView.tts?.stop();
-
-        if (requestPromise) {
-            requestPromise
-                .then((metrics) => this.requestCompleted(metrics))
-                .catch((error) => this.requestException(error));
-        }
-    }
-
-    public setMetricsVisible(visible: boolean) {
-        this.hideMetrics = !visible;
-        updateMetricsVisibility(visible, this.userMessageBody);
-        this.statusMessage?.setMetricsVisible(visible);
-        for (const agentMessage of this.agentMessages) {
-            agentMessage.setMetricsVisible(visible);
-        }
-    }
-    private requestCompleted(metrics: RequestMetrics | undefined) {
-        this.updateMetrics(metrics);
-        if (this.statusMessage === undefined) {
-            this.addStatusMessage(
-                { message: "Command completed", source: "shell" },
-                false,
-            );
-        } else {
-            this.statusMessage.complete();
-            this.chatView.updateScroll();
-        }
-    }
-
-    private requestException(error: any) {
-        console.error(error);
-        this.addStatusMessage(
-            { message: `Processing Error: ${error}`, source: "shell" },
-            false,
-        );
-    }
-
-    private ensureStatusMessage(source: string) {
-        if (this.statusMessage === undefined) {
-            this.statusMessage = new MessageContainer(
-                this.chatView,
-                "chat-message-left",
-                source,
-                this.agents,
-                this.userMessageContainer,
-                this.hideMetrics,
-                this.start,
-                true,
-            );
-        }
-
-        return this.statusMessage;
-    }
-
-    public addStatusMessage(msg: IAgentMessage, temporary: boolean) {
-        let message = msg.message;
-        const statusMessage = this.ensureStatusMessage(msg.source);
-        statusMessage.setMessage(
-            message,
-            msg.source,
-            temporary ? "temporary" : "block",
-        );
-
-        this.updateMetrics(msg.metrics);
-        this.chatView.updateScroll();
-    }
-
-    public updateMetrics(metrics?: RequestMetrics) {
-        if (metrics) {
-            if (metrics.parse !== undefined) {
-                if (this.metricsDiv === undefined) {
-                    const metricsContainer = document.createElement("div");
-                    metricsContainer.className =
-                        "chat-message-metrics chat-message-metrics-right";
-                    this.userMessageBody.append(metricsContainer);
-
-                    const metricsDetails = document.createElement("div");
-                    metricsDetails.className = "metrics-details";
-                    metricsContainer.append(metricsDetails);
-
-                    const left = document.createElement("div");
-                    metricsDetails.append(left);
-                    const right = document.createElement("div");
-                    metricsDetails.append(right);
-                    this.metricsDiv = {
-                        mainMetricsDiv: right,
-                        markMetricsDiv: left,
-                    };
-                }
-                updateMetrics(
-                    this.metricsDiv.mainMetricsDiv,
-                    this.metricsDiv.markMetricsDiv,
-                    "Translation",
-                    metrics.parse,
-                );
-            }
-
-            this.statusMessage?.updateMainMetrics(
-                metrics.command,
-                this.agentMessages.length === 0 ? metrics.duration : undefined,
-            );
-
-            for (let i = 0; i < this.agentMessages.length; i++) {
-                const agentMessage = this.agentMessages[i];
-                const info = metrics.actions[i];
-                agentMessage.updateMainMetrics(
-                    info,
-                    i === this.agentMessages.length - 1
-                        ? metrics.duration
-                        : undefined,
-                );
-            }
-        }
-    }
-
-    public ensureAgentMessage(msg: IAgentMessage, notification = false) {
-        const statusMessage = this.ensureStatusMessage(msg.source);
-
-        const index = msg.actionIndex;
-        if (index === undefined) {
-            return statusMessage;
-        }
-        const agentMessage = this.agentMessages[index];
-        if (agentMessage === undefined) {
-            statusMessage.setFirstResponseMetricsVisibility(false);
-            let beforeElem = statusMessage;
-            for (let i = 0; i < index + 1; i++) {
-                if (this.agentMessages[i] === undefined) {
-                    const newAgentMessage = new MessageContainer(
-                        this.chatView,
-                        "chat-message-left",
-                        msg.source,
-                        this.agents,
-                        beforeElem.div,
-                        this.hideMetrics,
-                        this.start,
-                        i === 0,
-                    );
-                    if (notification) {
-                        newAgentMessage.div.classList.add("notification");
-                    }
-                    this.agentMessages[i] = newAgentMessage;
-                }
-                beforeElem = this.agentMessages[i];
-            }
-            this.chatView.updateScroll();
-        }
-
-        this.updateMetrics(msg.metrics);
-        return this.agentMessages[index];
-    }
-
-    public updateMessageText(message: string) {
-        this.userMessage.textContent = message;
-    }
-}
-
-const ansi_up = new AnsiUp();
-ansi_up.use_classes = true;
-
-function textToHtml(text: string): string {
-    const value = ansi_up.ansi_to_html(text);
-    const line = value.replace(/\n/gm, "<br>");
-    return line;
-}
-
-function stripAnsi(text: string): string {
-    return text.replace(/\x1b\[[0-9;]*m/g, "");
-}
-
-function encodeTextToHtml(text: string): string {
-    return text
-        .replace(/&/gm, "&amp;")
-        .replace(/</gm, "&lt;")
-        .replace(/>/gm, "&gt;");
-}
-
-const enableText2Html = true;
-
-function matchKindStyle(elm: HTMLElement, kindStyle?: string) {
-    if (kindStyle !== undefined) {
-        return elm.classList.contains(kindStyle);
-    }
-    for (const cls of elm.classList) {
-        if (cls.startsWith("chat-message-kind-")) {
-            return false;
-        }
-    }
-    return true;
-}
-
-export function setContent(
-    elm: HTMLElement,
-    content: DisplayContent,
-    appendMode?: DisplayAppendMode,
-): string | undefined {
-    // Remove existing content if we are not appending.
-    if (appendMode === undefined) {
-        while (elm.firstChild) {
-            elm.removeChild(elm.firstChild);
-        }
-    }
-
-    let type: DisplayType;
-    let kind: DisplayMessageKind | undefined;
-    let text: string;
-    let speak: boolean;
-    let script: string | undefined;
-    if (typeof content === "string") {
-        type = "text";
-        text = content;
-        speak = false;
-        script = undefined;
-    } else {
-        type = content.type;
-        text = content.content;
-        kind = content.kind;
-        speak = content.speak ?? false;
-        script = content.script;
-    }
-
-    const kindStyle = kind ? `chat-message-kind-${kind}` : undefined;
-
-    let contentDiv = elm.lastChild as HTMLDivElement | null;
-    let newDiv = true;
-    if (
-        appendMode !== "inline" ||
-        !contentDiv ||
-        !matchKindStyle(contentDiv, kindStyle)
-    ) {
-        // Create a new div
-        contentDiv = document.createElement("div");
-        if (kindStyle) {
-            contentDiv.classList.add(kindStyle);
-        }
-        if (appendMode === "inline") {
-            contentDiv.style.display = "inline-block";
-        }
-        elm.appendChild(contentDiv);
-        newDiv = false;
-    }
-
-    let contentElm: HTMLElement = contentDiv;
-    if (type === "text") {
-        const prevElm = contentDiv.lastChild as HTMLElement | null;
-        if (prevElm?.classList.contains("chat-message-agent-text")) {
-            // If there is an existing text element then append to it.
-            contentElm = prevElm;
-        } else {
-            const span = document.createElement("span");
-            // create a text span so we can set "whitespace: break-spaces" css style of text content.
-            span.className = `chat-message-agent-text`;
-            contentDiv.appendChild(span);
-            contentElm = span;
-        }
-    }
-
-    // Process content according to type
-    const contentHtml =
-        type === "html"
-            ? DOMPurify.sanitize(text, {
-                  ADD_ATTR: ["target", "onclick", "onerror"],
-                  ADD_DATA_URI_TAGS: ["img"],
-                  ADD_URI_SAFE_ATTR: ["src"],
-              })
-            : enableText2Html
-              ? textToHtml(text)
-              : stripAnsi(encodeTextToHtml(text));
-
-    contentElm.innerHTML += contentHtml;
-
-    if (script) {
-        const scriptBlock: HTMLScriptElement = document.createElement("script");
-        scriptBlock.attributes["type"] = "text/javascript";
-        scriptBlock.innerHTML = script;
-
-        document.body.appendChild(scriptBlock);
-    }
-
-    if (!speak) {
-        return undefined;
-    }
-    if (type === "text") {
-        return stripAnsi(text);
-    }
-
-    if (newDiv) {
-        return contentElm.innerText;
-    }
-
-    const parser = new DOMParser();
-    return parser.parseFromString(contentHtml, "text/html").body.innerText;
-}
-
-export function createTimestampDiv(timestamp: Date, className: string) {
-    const timeStampDiv = document.createElement("div");
-    timeStampDiv.classList.add(className);
-
-    const nameSpan = document.createElement("span");
-    nameSpan.className = "agent-name";
-    timeStampDiv.appendChild(nameSpan); // name placeholder
-
-    const dateSpan = document.createElement("span");
-    dateSpan.className = "timestring";
-    timeStampDiv.appendChild(dateSpan); // time string
-
-    dateSpan.innerText = timestamp.toLocaleTimeString();
-
-    return timeStampDiv;
-}
-
-function updateMetrics(
-    mainMetricsDiv: HTMLDivElement,
-    markMetricsDiv: HTMLDivElement,
-    name: string,
-    metrics?: PhaseTiming,
-    total?: number,
-) {
-    // clear out previous perf data
-    mainMetricsDiv.innerHTML = "";
-    markMetricsDiv.innerHTML = "";
-
-    if (metrics?.marks) {
-        const messages: string[] = [];
-        for (const [key, value] of Object.entries(metrics.marks)) {
-            const { duration, count } = value;
-            messages.push(metricsString(key, duration, count));
-        }
-        markMetricsDiv.innerHTML = messages.join("<br>");
-    }
-    const messages: string[] = [];
-    if (metrics?.duration) {
-        messages.push(metricsString(`${name} Elapsed Time`, metrics.duration));
-    }
-    if (total !== undefined) {
-        messages.push(metricsString("Total Elapsed Time", total));
-    }
-    mainMetricsDiv.innerHTML = messages.join("<br>");
-}
-
-function formatTimeReaderFriendly(time: number) {
-    if (time < 1) {
-        return `${time.toFixed(3)}ms`;
-    } else if (time > 1000) {
-        return `${(time / 1000).toFixed(1)}s`;
-    } else {
-        return `${time.toFixed(1)}ms`;
     }
 }
